test_mbox_format: corrects regex substitution to match general cases
There might be cases where the diff line has a colon at the end of the line.
Correcting the regex so everything before the first occurrence of a colon
is substituted by an empty string. This will ensure the summary displays the
actual line that's causing the issue instead of empty strings.

Signed-off-by: Daniela Plascencia <daniela.plascencia@linux.intel.com>